+++ docs/conf.py
@@ -68,6 +68,7 @@ extensions = [
     'jupyter_sphinx.execute',
     'sphinx_autodoc_typehints',
     'reno.sphinxext',
+    'sphinx.ext.doctest',
 ]
 html_static_path = ['_static']
 templates_path = ['_templates']
@@ -80,6 +81,7 @@ html_css_files = [
 # -----------------------------------------------------------------------------
 
 autosummary_generate = True
+autosummary_generate_overwrite = False
 
 # -----------------------------------------------------------------------------
 # Autodoc
@@ -87,6 +89,7 @@ autosummary_generate = True
 
 autodoc_default_options = {
     'inherited-members': None,
+    'show-inheritance': None,
 }